requirements.txt
gradio
openai-whisper
gtts 
groq